requirements.txt
streamlit==1.40.0
openai==0.27.0
faiss-cpu==1.7.3
numpy==1.24.0
pandas==1.5.3
requests==2.31.0
python-dotenv
